removed/2020/CSAW-Quals/web/widthless/requirements.txt
click==7.1.2
Flask==1.1.2
flask-wtf
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
zwsp-steg-py==0.2.0
